order_service/requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
gunicorn==22.0.0
requests==2.32.3
# Pin to versions where SQS uses the query protocol (compatible with LocalStack)
boto3==1.28.63
botocore==1.31.63
PyJWT==2.8.0
